requirements.txt
fastapi~=0.112.2
uvicorn[standard]
kafka-python~=2.0.2
pymongo~=4.8.0
neo4j~=5.24.0
redis~=5.0.8

pydantic~=2.8.2